Change packaging method
Use Python's setuptools to create a Python package for the project. This
allows for using pip to install the project in the Docker container.
Additionally, a `tiptabs` script is now available on the command lines.

A Makefile is added with a few simple tasks, such as 'docker' to make
the Docker image, 'test' to run the tests (which have been moved to
their own directory), and 'env' to create a virtual environment with all
dependencies installed.